- Worker list moved to sidebar
