+++ src/app/app.module.ts
@@ -66,7 +66,7 @@ import { AlertComponent } from './components/alert/alert.component';
 import { ReplaceToBrNewLineCharPipe } from './pipes/replace-to-br-new-line-char.pipe';
 import { ClustersWrapperComponent } from './components/clusters-wrapper/clusters-wrapper.component';
 import { WorkersListComponent } from './components/clusters-wrapper/workers-list/workers-list.component';
-import { WorkerComponent } from './components/clusters-wrapper/workers-list/worker/worker.component';
+import { WorkerComponent } from './components/clusters-wrapper/worker/worker.component';
 
 @NgModule({
   declarations: [
