+++ src/app/components/clusters-wrapper/clusters-wrapper.component.html
@@ -1,1 +1,8 @@
-<router-outlet></router-outlet>
+<main class='mist-layout cs-height-100' fxLayout="row" fxLayoutAlign="space-between stretch">
+  <aside fxFlex="25%" class='endpoint-list'>
+    <mist-workers-list></mist-workers-list>
+  </aside>
+  <div fxFlex="75%" class='mdl-layout__content'>
+    <router-outlet></router-outlet>
+  </div>
+</main>
