+++ src/app/components/clusters-wrapper/worker/worker.component.scss
