+++ src/app/components/clusters-wrapper/workers-list/workers-list.component.scss
@@ -1,3 +1,12 @@
+@import "variables.scss";
 .workers-table {
   width: 100%;
 }
+
+.mdl-badge:after {
+    background-color: #f28a03;
+    height: $badge-width;
+    width: $badge-height;
+    font-size: 14px;
+    margin-right: 25px;
+  }
